+++ CHANGELOG.md
@@ -18,6 +18,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Security
 
+
+## 0.2.0 - Unreleased
+
+### Changed
+
+- Implement `CDCAggDatabase._prepare_validation_schema()`, which
+  returns the validation schema for Study record.
+  [Implements #14](https://bitbucket.org/cessda/cessda.cdc.aggregator.doc-store/issues/14)
+- Require latest commit of Kuha Document Store master
+  [Implements #14](https://bitbucket.org/cessda/cessda.cdc.aggregator.doc-store/issues/14)
+
+
 ## 0.1.0 - 2021-09-21
 
 ### Added